修复音频数据播放不流畅的问题。

+++ clickerpluslibrary/src/main/java/com/qcymall/clickerpluslibrary/utils/BLECMDUtil.kt
@@ -113,7 +113,7 @@ object BLECMDUtil {
         var result = 0
         if (d != null) {
             for (b in d) {
-                result = (result shl 8) + (b and 0xff.toByte())
+                result = (result shl 8) + (b.toInt() and 0xff)
             }
         }
         return result
